+++ src/components/Products.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import noImage from "../img/no-image.jpg";
 import { getProducts } from "../axios-services";
-import SingleProduct from "./SingleProduct";
 import '../style/Products.css'
+import Product from "./Product";
 
 
 const Products = ({ activeOrder, products, cartItems, setCartItems }) => {
@@ -43,7 +43,7 @@ const Products = ({ activeOrder, products, cartItems, setCartItems }) => {
 			<div className="products-container">
 				{products.map((product) => {
 					return (
-						<SingleProduct
+						<Product
 							key={product.id}
 							product={product}
 							activeOrder={activeOrder}
